Import patches for enabling i386 (#102)
patch-agpsupport:
Move linuxkpi/dummy/include/linux/agp_backend.h to
linuxkpi/gplv2/include/linux/agp_backend.h and define struct
agp_kern_info so we can use this instead of FreeBSD stuct agp_info.
In drm_agp_init copy data from agp_info to agp_kern_info.  Wherever
agp_info was used, make the code unmodified from Linux again.  This
fixes the problem where FreeBSD ai_aperture_size is in bytes and
Linux aper_size is in MiB.

Tell the VM subsystem that addresses in the AGP aperture are valid.  The
AGP device provides a range of addresses that are aliases for physical
memory.  The GPU uses this to access system RAM.  The CPU also has to
use these alias addresses instead of the real physical addresses so the
CPU caches work correctly when the CPU sees the GPU access these
addresses.

patch-drivers-gpu-drm-drm_bufs.c:
Fix a compiler error about incompatible pointers on i386.  The second
argument of vm_mmap has type vm_offset_t* which is unsigned int* on
i386 and not unsigned long*.

patch-drivers-gpu-drm-drm_os_config.h:
Only define COMPAT_FREEBSD32 on amd64.

patch-drivers-gpu-drm-drm_os_freebsd.h:
Add non-atomic definitions of readq and writeq like drm_os_linux.h on
Linux.  These are used in the i915 driver.

patch-drivers-gpu-drm-drm_vblank.c:
Export drm_timestamp_precision and drm_vblank_offdelay.  They are also
used in drm_sysctl_freebsd.c.

patch-drivers-gpu-drm-i915-i915_gem.c-1:
Fix a compiler error about incompatible pointers on i386.  The third
argument of invalidate_mapping_pages has type pgoff_t (unsigned long),
not loff_t (unsigned long long on i386).  This should probably be
upstreamed.

patch-drivers-gpu-drm-i915-i915_gem.c-2:
Fix a compiler error about incompatible pointers on i386.  The fourth
argument of vm_map_find has type vm_offset_t* which is unsigned int* on
i386 and not unsigned long*.

patch-drivers-gpu-drm-radeon-radeon_drv.c:
Default radeon_hw_i2c back to 0 like Linux because the i2c support on
older cards appears to be broken.  This reverts
https://github.com/FreeBSDDesktop/DEPRECATED-freebsd-base-graphics/commit/eecaad151b81f6ef1809ee274d02bef52460e755
jkim said he no longer needs it:
https://github.com/FreeBSDDesktop/kms-drm/issues/19#issuecomment-376715938

Maybe amdgpu_hw_i2c can be set back to zero as well.

patch-linuxkpi-gplv2-include-asm-fpu-api.h:
fpu.h is called npx.h on i386 (for historic reasons).

patch-linuxkpi-gplv2-src-linux_notifier.c:
Use atomic_fetchadd_long instead of atomic_fetchadd_64 because the variable
operated on is long which is 32 bit on i386.

patch-linuxkpi-gplv2-src-linux_page.c:
Include sys/sf_buf.h for architectures like i386 without DMAP.

patch-linuxkpi-gplv2-src-linux_sync_file.c:
Print uint64_t value using uintmax_t because it's unsigned long on amd64
and unsigned long long on i386.

+++ linuxkpi/gplv2/src/linux_sync_file.c
@@ -144,10 +144,10 @@ char *sync_file_get_name(struct sync_file *sync_file, char *buf, int len)
 	} else {
 		struct dma_fence *fence = sync_file->fence;
 
-		snprintf(buf, len, "%s-%s%lu-%d",
+		snprintf(buf, len, "%s-%s%ju-%d",
 			 fence->ops->get_driver_name(fence),
 			 fence->ops->get_timeline_name(fence),
-			 fence->context,
+			 (uintmax_t)fence->context,
 			 fence->seqno);
 	}
